32bit: Fix wrong format string

+++ lib/hardware.c
@@ -1454,7 +1454,8 @@ int h_inflate(z_streamp strm, int flush)
 			rc = __check_stream_end(strm);
 			if (rc == Z_STREAM_END) {
 				hw_trace("    Suppress Z_STREAM_END %ld %ld\n",
-					 s->obuf_avail, s->obuf_total);
+					 (unsigned long)s->obuf_avail,
+					 (unsigned long)s->obuf_total);
 				s->rc = Z_STREAM_END;
 				rc = Z_OK;
 			}
